refactor file structure
condense sign.js and verify.js into single cryptosuite creation
