+++ lib/index.js
@@ -1,5 +1,4 @@
 /*!
  * Copyright (c) 2024 Digital Bazaar, Inc. All rights reserved.
 */
-export {createSignCryptosuite} from './sign.js';
-export {createVerifyCryptosuite} from './verify.js';
+export {createCryptosuite} from './create.js';
