+++ lib/test.js
@@ -1,7 +1,7 @@
 import * as EcdsaMultikey from '@digitalbazaar/ecdsa-multikey';
 import {DataIntegrityProof} from '@digitalbazaar/data-integrity';
 import jsigs from 'jsonld-signatures';
-import { createCryptosuite } from './sign.js';
+import { createCryptosuite } from './create.js';
 const {purposes: {AssertionProofPurpose}} = jsigs;
 import {ecdsaMultikeyKeyPair} from '../test/mock-data.js'
 import {loader} from '../test/documentLoader.js';
